# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Documentation",
    "version": "1.0.0",
    "description": "This is the API documentation for the Employee Directory and Branch Management applciation."
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Local Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Branch": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier for the branch"
          },
          "name": {
            "type": "string",
            "description": "The branch name"
          },
          "location": {
            "type": "string",
            "description": "Physical address or location of the branch"
          }
        },
        "required": [
          "id",
          "name"
        ]
      },
      "Employee": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier for the employee"
          },
          "firstName": {
            "type": "string",
            "description": "Employee's first name"
          },
          "lastName": {
            "type": "string",
            "description": "Employee's last name"
          },
          "department": {
            "type": "string",
            "description": "Department where the employee works"
          },
          "branchId": {
            "type": "number",
            "description": "ID of the branch the employee is assigned to"
          }
        },
        "required": [
          "id",
          "firstName",
          "lastName",
          "department"
        ]
      },
      "Item": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier for the item"
          },
          "name": {
            "type": "string",
            "description": "Name of the item"
          },
          "description": {
            "type": "string",
            "description": "Description of the item"
          },
          "price": {
            "type": "number",
            "description": "Price of the item (optional)"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "When the item was created"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "When the item was last updated"
          }
        },
        "required": [
          "id",
          "name",
          "description"
        ]
      }
    }
  },
  "paths": {
    "/branches": {
      "get": {
        "summary": "Get all branches",
        "tags": [
          "Branches"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "A list of branches",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Branch"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new branch",
        "tags": [
          "Branches"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "description": "Branch data to create",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Branch"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Branch created successfully"
          },
          "400": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/branches/{id}": {
      "get": {
        "summary": "Get a branch by ID",
        "tags": [
          "Branches"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The ID of the branch"
          }
        ],
        "responses": {
          "200": {
            "description": "Branch details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Branch"
                }
              }
            }
          },
          "404": {
            "description": "Branch not found"
          }
        }
      },
      "delete": {
        "summary": "Delete a branch",
        "tags": [
          "Branches"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The ID of the branch"
          }
        ],
        "responses": {
          "200": {
            "description": "Branch deleted successfully"
          },
          "404": {
            "description": "Branch not found"
          }
        }
      },
      "put": {
        "summary": "Update an existing branch",
        "tags": [
          "Branches"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The ID of the branch"
          }
        ],
        "requestBody": {
          "description": "Updated branch data",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Branch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Branch updated successfully"
          },
          "404": {
            "description": "Branch not found"
          }
        }
      }
    },
    "/employees": {
      "get": {
        "summary": "Get all employees",
        "tags": [
          "Employees"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "A list of employees",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Employee"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new employee",
        "tags": [
          "Employees"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "description": "Employee data to create",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Employee"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Employee created"
          },
          "400": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/employees/branch/{branchId}": {
      "get": {
        "summary": "Get employees for a specific branch",
        "tags": [
          "Employees"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "branchId",
            "required": true,
            "schema": {
              "type": "number"
            },
            "description": "The ID of the branch"
          }
        ],
        "responses": {
          "200": {
            "description": "A list of employees for the given branch"
          }
        }
      }
    },
    "/employees/department/{department}": {
      "get": {
        "summary": "Get employees by department",
        "tags": [
          "Employees"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "department",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The department name"
          }
        ],
        "responses": {
          "200": {
            "description": "A list of employees in the specified department"
          }
        }
      }
    },
    "/employees/{id}": {
      "get": {
        "summary": "Get an employee by ID",
        "tags": [
          "Employees"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The ID of the employee"
          }
        ],
        "responses": {
          "200": {
            "description": "An employee object",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Employee"
                }
              }
            }
          },
          "404": {
            "description": "Employee not found"
          }
        }
      },
      "delete": {
        "summary": "Delete an employee",
        "tags": [
          "Employees"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The ID of the employee"
          }
        ],
        "responses": {
          "200": {
            "description": "Employee deleted successfully"
          },
          "404": {
            "description": "Employee not found"
          }
        }
      },
      "put": {
        "summary": "Update an existing employee",
        "tags": [
          "Employees"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The ID of the employee"
          }
        ],
        "requestBody": {
          "description": "Updated employee data",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Employee"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Employee updated"
          },
          "404": {
            "description": "Employee not found"
          }
        }
      }
    }
  },
  "tags": []
}